// src/grpc/proto/obsidian_events.ts
export class CreateEvent extends pb_1.Message {
	#one_of_decls: number[][] = [];
	constructor(
		data?:
			| any[]
			| {
					filePath?: string;
			  }
	) {
		super();
		pb_1.Message.initialize(
			this,
			Array.isArray(data) ? data : [],
			0,
			-1,
			[],
			this.#one_of_decls
		);
		if (!Array.isArray(data) && typeof data == "object") {
			if ("filePath" in data && data.filePath != undefined) {
				this.filePath = data.filePath;
			}
		}
	}
	get filePath() {
		return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
	}
	set filePath(value: string) {
		pb_1.Message.setField(this, 1, value);
	}
	static fromObject(data: { filePath?: string }): CreateEvent {
		const message = new CreateEvent({});
		if (data.filePath != null) {
			message.filePath = data.filePath;
		}
		return message;
	}
	toObject() {
		const data: {
			filePath?: string;
		} = {};
		if (this.filePath != null) {
			data.filePath = this.filePath;
		}
		return data;
	}
	serialize(): Uint8Array;
	serialize(w: pb_1.BinaryWriter): void;
	serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
		const writer = w || new pb_1.BinaryWriter();
		if (this.filePath.length) writer.writeString(1, this.filePath);
		if (!w) return writer.getResultBuffer();
	}
	static deserialize(bytes: Uint8Array | pb_1.BinaryReader): CreateEvent {
		const reader =
				bytes instanceof pb_1.BinaryReader
					? bytes
					: new pb_1.BinaryReader(bytes),
			message = new CreateEvent();
		while (reader.nextField()) {
			if (reader.isEndGroup()) break;
			switch (reader.getFieldNumber()) {
				case 1:
					message.filePath = reader.readString();
					break;
				default:
					reader.skipField();
			}
		}
		return message;
	}
	serializeBinary(): Uint8Array {
		return this.serialize();
	}
	static deserializeBinary(bytes: Uint8Array): CreateEvent {
		return CreateEvent.deserialize(bytes);
	}
}
export class DeleteEvent extends pb_1.Message {
	#one_of_decls: number[][] = [];
	constructor(
		data?:
			| any[]
			| {
					filePath?: string;
			  }
	) {
		super();
		pb_1.Message.initialize(
			this,
			Array.isArray(data) ? data : [],
			0,
			-1,
			[],
			this.#one_of_decls
		);
		if (!Array.isArray(data) && typeof data == "object") {
			if ("filePath" in data && data.filePath != undefined) {
				this.filePath = data.filePath;
			}
		}
	}
	get filePath() {
		return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
	}
	set filePath(value: string) {
		pb_1.Message.setField(this, 1, value);
	}
	static fromObject(data: { filePath?: string }): DeleteEvent {
		const message = new DeleteEvent({});
		if (data.filePath != null) {
			message.filePath = data.filePath;
		}
		return message;
	}
	toObject() {
		const data: {
			filePath?: string;
		} = {};
		if (this.filePath != null) {
			data.filePath = this.filePath;
		}
		return data;
	}
	serialize(): Uint8Array;
	serialize(w: pb_1.BinaryWriter): void;
	serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
		const writer = w || new pb_1.BinaryWriter();
		if (this.filePath.length) writer.writeString(1, this.filePath);
		if (!w) return writer.getResultBuffer();
	}
	static deserialize(bytes: Uint8Array | pb_1.BinaryReader): DeleteEvent {
		const reader =
				bytes instanceof pb_1.BinaryReader
					? bytes
					: new pb_1.BinaryReader(bytes),
			message = new DeleteEvent();
		while (reader.nextField()) {
			if (reader.isEndGroup()) break;
			switch (reader.getFieldNumber()) {
				case 1:
					message.filePath = reader.readString();
					break;
				default:
					reader.skipField();
			}
		}
		return message;
	}
	serializeBinary(): Uint8Array {
		return this.serialize();
	}
	static deserializeBinary(bytes: Uint8Array): DeleteEvent {
		return DeleteEvent.deserialize(bytes);
	}
}
export class RenameEvent extends pb_1.Message {
	#one_of_decls: number[][] = [];
	constructor(
		data?:
			| any[]
			| {
					newPath?: string;
					oldPath?: string;
			  }
	) {
		super();
		pb_1.Message.initialize(
			this,
			Array.isArray(data) ? data : [],
			0,
			-1,
			[],
			this.#one_of_decls
		);
		if (!Array.isArray(data) && typeof data == "object") {
			if ("newPath" in data && data.newPath != undefined) {
				this.newPath = data.newPath;
			}
			if ("oldPath" in data && data.oldPath != undefined) {
				this.oldPath = data.oldPath;
			}
		}
	}
	get newPath() {
		return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
	}
	set newPath(value: string) {
		pb_1.Message.setField(this, 1, value);
	}
	get oldPath() {
		return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
	}
	set oldPath(value: string) {
		pb_1.Message.setField(this, 2, value);
	}
	static fromObject(data: {
		newPath?: string;
		oldPath?: string;
	}): RenameEvent {
		const message = new RenameEvent({});
		if (data.newPath != null) {
			message.newPath = data.newPath;
		}
		if (data.oldPath != null) {
			message.oldPath = data.oldPath;
		}
		return message;
	}
	toObject() {
		const data: {
			newPath?: string;
			oldPath?: string;
		} = {};
		if (this.newPath != null) {
			data.newPath = this.newPath;
		}
		if (this.oldPath != null) {
			data.oldPath = this.oldPath;
		}
		return data;
	}
	serialize(): Uint8Array;
	serialize(w: pb_1.BinaryWriter): void;
	serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
		const writer = w || new pb_1.BinaryWriter();
		if (this.newPath.length) writer.writeString(1, this.newPath);
		if (this.oldPath.length) writer.writeString(2, this.oldPath);
		if (!w) return writer.getResultBuffer();
	}
	static deserialize(bytes: Uint8Array | pb_1.BinaryReader): RenameEvent {
		const reader =
				bytes instanceof pb_1.BinaryReader
					? bytes
					: new pb_1.BinaryReader(bytes),
			message = new RenameEvent();
		while (reader.nextField()) {
			if (reader.isEndGroup()) break;
			switch (reader.getFieldNumber()) {
				case 1:
					message.newPath = reader.readString();
					break;
				case 2:
					message.oldPath = reader.readString();
					break;
				default:
					reader.skipField();
			}
		}
		return message;
	}
	serializeBinary(): Uint8Array {
		return this.serialize();
	}
	static deserializeBinary(bytes: Uint8Array): RenameEvent {
		return RenameEvent.deserialize(bytes);
	}
}
export class ModifyEvent extends pb_1.Message {
	#one_of_decls: number[][] = [];
	constructor(
		data?:
			| any[]
			| {
					filePath?: string;
			  }
	) {
		super();
		pb_1.Message.initialize(
			this,
			Array.isArray(data) ? data : [],
			0,
			-1,
			[],
			this.#one_of_decls
		);
		if (!Array.isArray(data) && typeof data == "object") {
			if ("filePath" in data && data.filePath != undefined) {
				this.filePath = data.filePath;
			}
		}
	}
	get filePath() {
		return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
	}
	set filePath(value: string) {
		pb_1.Message.setField(this, 1, value);
	}
	static fromObject(data: { filePath?: string }): ModifyEvent {
		const message = new ModifyEvent({});
		if (data.filePath != null) {
			message.filePath = data.filePath;
		}
		return message;
	}
	toObject() {
		const data: {
			filePath?: string;
		} = {};
		if (this.filePath != null) {
			data.filePath = this.filePath;
		}
		return data;
	}
	serialize(): Uint8Array;
	serialize(w: pb_1.BinaryWriter): void;
	serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
		const writer = w || new pb_1.BinaryWriter();
		if (this.filePath.length) writer.writeString(1, this.filePath);
		if (!w) return writer.getResultBuffer();
	}
	static deserialize(bytes: Uint8Array | pb_1.BinaryReader): ModifyEvent {
		const reader =
				bytes instanceof pb_1.BinaryReader
					? bytes
					: new pb_1.BinaryReader(bytes),
			message = new ModifyEvent();
		while (reader.nextField()) {
			if (reader.isEndGroup()) break;
			switch (reader.getFieldNumber()) {
				case 1:
					message.filePath = reader.readString();
					break;
				default:
					reader.skipField();
			}
		}
		return message;
	}
	serializeBinary(): Uint8Array {
		return this.serialize();
	}
	static deserializeBinary(bytes: Uint8Array): ModifyEvent {
		return ModifyEvent.deserialize(bytes);
	}
}
export class FileOpenEvent extends pb_1.Message {
	#one_of_decls: number[][] = [];
	constructor(
		data?:
			| any[]
			| {
					filePath?: string;
			  }
	) {
		super();
		pb_1.Message.initialize(
			this,
			Array.isArray(data) ? data : [],
			0,
			-1,
			[],
			this.#one_of_decls
		);
		if (!Array.isArray(data) && typeof data == "object") {
			if ("filePath" in data && data.filePath != undefined) {
				this.filePath = data.filePath;
			}
		}
	}
	get filePath() {
		return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
	}
	set filePath(value: string) {
		pb_1.Message.setField(this, 1, value);
	}
	static fromObject(data: { filePath?: string }): FileOpenEvent {
		const message = new FileOpenEvent({});
		if (data.filePath != null) {
			message.filePath = data.filePath;
		}
		return message;
	}
	toObject() {
		const data: {
			filePath?: string;
		} = {};
		if (this.filePath != null) {
			data.filePath = this.filePath;
		}
		return data;
	}
	serialize(): Uint8Array;
	serialize(w: pb_1.BinaryWriter): void;
	serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
		const writer = w || new pb_1.BinaryWriter();
		if (this.filePath.length) writer.writeString(1, this.filePath);
		if (!w) return writer.getResultBuffer();
	}
	static deserialize(bytes: Uint8Array | pb_1.BinaryReader): FileOpenEvent {
		const reader =
				bytes instanceof pb_1.BinaryReader
					? bytes
					: new pb_1.BinaryReader(bytes),
			message = new FileOpenEvent();
		while (reader.nextField()) {
			if (reader.isEndGroup()) break;
			switch (reader.getFieldNumber()) {
				case 1:
					message.filePath = reader.readString();
					break;
				default:
					reader.skipField();
			}
		}
		return message;
	}
	serializeBinary(): Uint8Array {
		return this.serialize();
	}
	static deserializeBinary(bytes: Uint8Array): FileOpenEvent {
		return FileOpenEvent.deserialize(bytes);
	}
}
// ...
interface GrpWritableServiceInterface<P, R> {
	(
		metadata: grpc_1.Metadata,
		options: grpc_1.CallOptions,
		callback: grpc_1.requestCallback<R>
	): grpc_1.ClientWritableStream<P>;
	(
		metadata: grpc_1.Metadata,
		callback: grpc_1.requestCallback<R>
	): grpc_1.ClientWritableStream<P>;
	(
		options: grpc_1.CallOptions,
		callback: grpc_1.requestCallback<R>
	): grpc_1.ClientWritableStream<P>;
	(callback: grpc_1.requestCallback<R>): grpc_1.ClientWritableStream<P>;
}
// ...
export abstract class UnimplementedObsidianEventStreamService {
	static definition = {
		streamEvents: {
			path: "/ObsidianEventStream/streamEvents",
			requestStream: true,
			responseStream: false,
			requestSerialize: (message: ObsidianEvent) =>
				Buffer.from(message.serialize()),
			requestDeserialize: (bytes: Buffer) =>
				ObsidianEvent.deserialize(new Uint8Array(bytes)),
			responseSerialize: (message: Empty) =>
				Buffer.from(message.serialize()),
			responseDeserialize: (bytes: Buffer) =>
				Empty.deserialize(new Uint8Array(bytes)),
		},
	};
	[method: string]: grpc_1.UntypedHandleCall;
	abstract streamEvents(
		call: grpc_1.ServerReadableStream<ObsidianEvent, Empty>,
		callback: grpc_1.sendUnaryData<Empty>
	): void;
}
export class ObsidianEventStreamClient extends grpc_1.makeGenericClientConstructor(
	UnimplementedObsidianEventStreamService.definition,
	"ObsidianEventStream",
	{}
) {
	constructor(
		address: string,
		credentials: grpc_1.ChannelCredentials,
		options?: Partial<grpc_1.ChannelOptions>
	) {
		super(address, credentials, options);
	}
	streamEvents: GrpWritableServiceInterface<ObsidianEvent, Empty> = (
		metadata:
			| grpc_1.Metadata
			| grpc_1.CallOptions
			| grpc_1.requestCallback<Empty>,
		options?: grpc_1.CallOptions | grpc_1.requestCallback<Empty>,
		callback?: grpc_1.requestCallback<Empty>
	): grpc_1.ClientWritableStream<ObsidianEvent> => {
		return super.streamEvents(metadata, options, callback);
	};
}